Move the entire IntroModal banner component into a new notifications folder.
