+++ assets/js/modules/analytics-4/components/site-goals/notifications/IntroModalBanner/IntroModalEcommerceAndLead.tsx
@@ -29,7 +29,7 @@ import BannerModal from '@/js/components/BannerModal/index';
 // @ts-expect-error - We need to add types for imported SVGs.
 import SiteGoalsIntroModalGraphic from '@/svg/graphics/site-goals-intro-modal.svg';
 import Link from '@/js/components/Link';
-import { IntroModalVariantProps } from '@/js/modules/analytics-4/components/site-goals/IntroModalBanner/index';
+import { IntroModalVariantProps } from '@/js/modules/analytics-4/components/site-goals/notifications/IntroModalBanner/index';
 
 export default function IntroModalEcommerceAndLead( {
 	onView,
